feat: add pipelines and initial context
Add new Pipeline and PipelineStage classes that enable flexible and modular programming. The idea is
to mimic the modules of a video playback pipeline with the ability to swap in and out stages based
on desired functionality. These can be removed and replaced with a single function as
experimentation dies down and code becomes stable.